Review OpenFile#read methods
Update the OpenFile#read method to avoid to reach EOF
and try to read the maximum of possible bytes.

To do that, re-evaluate the num value by choosing
the min value between the given num and the
actual file channel size minus the given offset.

In that way it is not possible to exceed the EOF and
number of bytes read returned is always correct.

Signed-off-by: Callimard <guil.rako@hotmail.fr>